+++ flytekit/clis/sdk_in_container/backfill.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 import rich_click as click
 
 from flytekit.clis.sdk_in_container.helpers import get_and_save_remote_with_click_context
-from flytekit.clis.sdk_in_container.run import DateTimeType, DurationParamType
+from flytekit.interaction.click_types import DateTimeType, DurationParamType
 
 _backfill_help = """
 The backfill command generates and registers a new workflow based on the input launchplan to run an
